autopay_web/autopay_web_v42/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
plotly
openpyxl
oracledb
python-dotenv

requests>=2.31.0
